requirements.txt
click==6.7
decorator==4.0.11
imageio==2.1.2
matplotlib==2.0.2
moviepy==0.2.3.2
numpy==1.13.1
olefile==0.44
packaging==16.8
pbr==3.0.0
Pillow==4.2.1
pycairo==1.13.3
pysrt==1.1.1
pytz==2017.2
tqdm==4.11.2